python_app/requirement.txt
setuptools 72.1.0
hidapi==0.14.0.post2
setuptools==72.1.0
